fix and improve progress bar

+++ src/create-loc-history.hpp
@@ -85,7 +85,7 @@ class Commit {
 
 std::vector<Commit> create_loc_history(
     std::string git_repo_path, std::vector<std::string> excluded_paths,
-    std::function<void(double, clock_t)> on_progress,
+    std::function<void(int, clock_t)> on_progress,
     std::function<void(std::string, clock_t)> on_section_change,
     const std::clock_t start
 );
